scpi_psu: Move scpi_psu agent into package
