+++ agents/ocs_plugin_so.py
@@ -9,7 +9,6 @@ import os
 root = os.path.abspath(os.path.split(__file__)[0])
 
 for n, f in [
-        ('Keithley2230G-PSU', 'keithley2230G-psu/keithley_agent.py'),
         ('PysmurfController', 'pysmurf_controller/pysmurf_controller.py'),
         ('LATRtXYStageAgent', 'xy_stage/xy_latrt_agent.py'),
         ('VantagePro2Agent', 'vantagePro2_agent/vantage_pro2_agent.py'),
