+++ docs/api.rst
@@ -32,14 +32,6 @@ socs.agent.prologix_interface
     :undoc-members:
     :show-inheritance:
 
-socs.agent.scpi_psu_driver
-``````````````````````````
-
-.. automodule:: socs.agent.scpi_psu_driver
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
 socs.agent.tektronix3021c_driver
 ````````````````````````````````
 
@@ -52,6 +44,17 @@ socs.agent.vantage_pro2
 ```````````````````````
 
 .. automodule:: socs.agent.vantage_pro2.vantage_pro2
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+socs.agents
+-----------
+
+socs.agents.scpi_psu.drivers
+````````````````````````````
+
+.. automodule:: socs.agents.scpi_psu.drivers
     :members:
     :undoc-members:
     :show-inheritance:
