+++ tests/agents/test_scpi_psu_agent.py
@@ -1,3 +1,1 @@
-import sys
-sys.path.insert(0, '../agents/scpi_psu/')
-from scpi_psu_agent import ScpiPsuAgent  # noqa: F401
+from socs.agents.scpi_psu.agent import ScpiPsuAgent  # noqa: F401
