+++ tests/integration/test_scpi_psu_agent_integration.py
@@ -16,7 +16,7 @@ pytest_plugins = "docker_compose"
 
 wait_for_crossbar = create_crossbar_fixture()
 run_agent = create_agent_runner_fixture(
-    "../agents/scpi_psu/scpi_psu_agent.py",
+    "../socs/agents/scpi_psu/agent.py",
     "scpi_psu_agent",
     args=["--log-dir", "./logs/"],
 )
